Changed code placement for EventEmitter
