+++ llrt_core/src/modules/net/socket.rs
@@ -28,7 +28,7 @@ use tracing::trace;
 
 use crate::{
     modules::{
-        events::{event_emitter::EventEmitter, EmitError, Emitter, EventKey, EventList},
+        events::{EmitError, Emitter, EventEmitter, EventKey, EventList},
         module::export_default,
     },
     security::ensure_net_access,
